chore: Delete the custom http client
It is just adding complications
